Certain fonts and images can now also be served.
The task hasn't been modified yet. Images and fonts should be copied to
the static dir. I'm not entirely sure this is the job of the assets
task, as that would mean that either:
* a list of fonts/images has to be put in the manifest; or
* the css has to be checked for references to fonts/images.
I'd say if you have assets outside of stylesheets and javascripts, you
should copy these files yourself.